requirements.txt
openapi_spec_validator
pyyaml
ruamel.yaml
jsonpath_ng